requirements.txt
# Task Maistro - Railway Deployment Requirements (solo Postgres persistente, sin Redis)
# Solo dependencias esenciales para producción Railway

gradio==4.44.0
python-dotenv==1.0.0
langchain-core>=0.3.0
langchain-openai>=0.2.8 
langgraph[redis]>=0.2.50
langgraph-prebuilt
langgraph-sdk
langsmith
langchain-community
trustcall==0.0.20
pydantic==2.9.2
httpx==0.27.2
typing-extensions==4.12.2

